fix buggy env delete, wrong access info, clean up
Problem 1:
Deleting some enviroment variables, specifically variables that had similar
names with other variables, such as USER and USERNAME, resulted in the deletion
of both, causing an empty variable in the middle of the env array, essentially
losing most of the array.

Solution:
While searching for the specified enviroment variable, checking that the
name value matches exactly, and the following character of the name value
is a '=' character, indicating that the name has ended. Only one enviroment
variable is now deleted.

Problem 2:
Occasionally, the error "Permission denied" was displayed even though the
program was executable.

Solution:
Instead of checking if execution is possible with access, which can return
a multitude of errors, getting correct permission output has now been
achieved by checking if execution fails.

Problem 3:
If the enviroment variables USER or PWD were deleted, a segfault would occur.

Solution:
This was due to USER and PWD being printed on screen with each loop.
Now, if the USER enviroment variable does not exist, printing is not attempted.
PWD is not used anymore for printing the current directory, as this information
would not be available if PWD was deleted, even though it is information
that is always fetchable. Instead, the current directory is now fetched with
getcwd, and printed to the screen on each loop.

'~' expansion has been changed to only expand one tilde character per argument,
if several exist in a single argument.

Code has been cleaned up to comform to the Norm.